requirements.txt
aiohttp==3.8.1
async-timeout==4.0.2
asyncio==3.4.3
asyncpg==0.23.0
config==0.5.1
DateTime==4.4
dnspython==2.1.0
easy-pil==0.0.2
env==0.1.0
frozenlist==1.3.0
function-cooldowns==1.3.0
ipc==1.2.1
install==1.3.5
nextcord==2.0.0b4
nextcord-ext-menus==1.5.4
nextcord-ext-ipc==2.2.1
Pillow==8.4.0
pymongo==3.11.4
python-dotenv==0.17.1
requests==2.26.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
quart==0.17.0
Quart-Discord==2.1.4
gunicorn==20.0.4
uvicorn==0.18.2
lazy-object-proxy==1.4.3
pylint==2.6.0